Query RAG for each wine individually (#38)
Querying pinecone RAG for all the wines on the menu page did not result
in accurate search results. First, on pinecone itself, I separated notes
from the domaine into distinct records for the domaine and each of the
wines listed. Then, in this repo, I will now query each of the wines
returned from the menu page; sort the results, de-dupe them, and use the
best 10.

I also did a bit of prompt engineering because the AI was not always
recommending wines from the wine list.

The prompt to read the wine list image was already returning the wines
as a list; but perhaps in the future I could have this returned as a
JSON schema if the format is not consistently parseable.